Added additional Scan Chain stuff and changed the rng to be long long. Also, moved the polynomial storage into the rng, as it seems more natural.

+++ C++/FSMSynthesis/ErrorDetectionNetwork.h
@@ -36,8 +36,8 @@ public:
 
 	//Getters and Setters
 	void setType(ErrorDetectionNetworkType);
-	void setRandNumGenPolynomial(unsigned int polynomial);
-	unsigned int getPolynomial(int degree);
-	unsigned int getRandNumber();
-	void seedRNG(unsigned int seed);
+	void setRandNumGenPolynomial(unsigned long long polynomial);
+	unsigned long long getPolynomial(int degree);
+	unsigned long long getRandNumber();
+	void seedRNG(unsigned long long seed);
 };
